ci: add PR size check
